Reorganzied into a more maintainable way
