+++ application/src/main/java/com/github/pietergroenendijk/UserContactDetails.java
@@ -1,4 +1,4 @@
-package com.github.pietergroenendijk.notifications;
+package com.github.pietergroenendijk;
 
 public record UserContactDetails(String email, String phoneNumber) {
