--- application/src/main/java/com/github/pietergroenendijk/notifications/Notification.java
@@ -1,5 +1,0 @@
-package com.github.pietergroenendijk.notifications;
-
-public record Notification(String title, String message){
-
-}
