--- application/src/main/java/com/github/pietergroenendijk/notifications/NotificationTaskStoreStrategy.java
@@ -1,5 +1,0 @@
-package com.github.pietergroenendijk.notifications;
-
-public interface NotificationTaskStoreStrategy {
-    public void store(NotificationTask task);
-}
