--- application/src/main/java/com/github/pietergroenendijk/notifications/notifiers/AppNotifier.java
@@ -1,10 +1,0 @@
-package com.github.pietergroenendijk.notifications.notifiers;
-
-import com.github.pietergroenendijk.notifications.NotificationTask;
-
-public class AppNotifier extends ChannelNotifier {
-    @Override
-    protected void attempt(NotificationTask task) throws Exception {
-
-    }
-}
