--- application/src/main/java/com/github/pietergroenendijk/notifications/notifiers/EmailNotifier.java
@@ -1,9 +1,0 @@
-package com.github.pietergroenendijk.notifications.notifiers;
-
-import com.github.pietergroenendijk.notifications.NotificationTask;
-
-public class EmailNotifier extends ChannelNotifier {
-    @Override
-    protected void attempt(NotificationTask task) throws Exception {
-    }
-}
