--- application/src/main/java/com/github/pietergroenendijk/notifications/strategy/ReminderNotificationSendStrategy.java
@@ -1,12 +1,0 @@
-package com.github.pietergroenendijk.notifications.strategy;
-
-import com.github.pietergroenendijk.notifications.notifiers.EmailNotifier;
-import com.github.pietergroenendijk.notifications.notifiers.Notifier;
-
-public class ReminderNotificationSendStrategy extends NotificationSendStrategy {
-    public ReminderNotificationSendStrategy(EmailNotifier emailNotifier) {
-        super(new Notifier[]{
-            emailNotifier
-        });
-    }
-}
