+++ application/src/main/java/com/github/pietergroenendijk/services/notifications/notifiers/EmailNotifier.java
@@ -1,0 +1,9 @@
+package com.github.pietergroenendijk.services.notifications.notifiers;
+
+import com.github.pietergroenendijk.services.notifications.task.NotificationTask;
+
+public class EmailNotifier extends ChannelNotifier {
+    @Override
+    protected void attempt(NotificationTask task) throws Exception {
+    }
+}
