+++ application/src/main/java/com/github/pietergroenendijk/services/notifications/notifiers/Notifier.java
@@ -1,0 +1,7 @@
+package com.github.pietergroenendijk.services.notifications.notifiers;
+
+import com.github.pietergroenendijk.services.notifications.task.NotificationTask;
+
+public interface Notifier {
+    void send(NotificationTask task);
+}
