+++ application/src/main/java/com/github/pietergroenendijk/services/notifications/notifiers/SMSNotifier.java
@@ -1,0 +1,10 @@
+package com.github.pietergroenendijk.services.notifications.notifiers;
+
+import com.github.pietergroenendijk.services.notifications.task.NotificationTask;
+
+public class SMSNotifier extends ChannelNotifier {
+    @Override
+    protected void attempt(NotificationTask task) throws Exception {
+
+    }
+}
