+++ application/src/main/java/com/github/pietergroenendijk/services/notifications/send_strategies/ReminderNotificationSendStrategy.java
@@ -1,0 +1,12 @@
+package com.github.pietergroenendijk.services.notifications.send_strategies;
+
+import com.github.pietergroenendijk.services.notifications.notifiers.EmailNotifier;
+import com.github.pietergroenendijk.services.notifications.notifiers.Notifier;
+
+public class ReminderNotificationSendStrategy extends NotificationSendStrategy {
+    public ReminderNotificationSendStrategy(EmailNotifier emailNotifier) {
+        super(new Notifier[]{
+            emailNotifier
+        });
+    }
+}
