+++ application/src/main/java/com/github/pietergroenendijk/services/notifications/task/Notification.java
@@ -1,0 +1,5 @@
+package com.github.pietergroenendijk.services.notifications.task;
+
+public record Notification(String title, String message){
+
+}
