+++ application/src/main/java/com/github/pietergroenendijk/services/notifications/task/NotificationTaskFactory.java
@@ -1,13 +1,12 @@
-package com.github.pietergroenendijk.notifications.generators;
+package com.github.pietergroenendijk.services.notifications.task;
 
 import com.github.pietergroenendijk.AccountBase;
 import com.github.pietergroenendijk.Lending;
-import com.github.pietergroenendijk.notifications.NotificationTask;
-import com.github.pietergroenendijk.notifications.NotificationTaskRepository;
-import com.github.pietergroenendijk.notifications.generators.returndate.ReturnDateNotificationTaskGenerator;
-import com.github.pietergroenendijk.notifications.strategy.AlertNotificationSendStrategy;
-import com.github.pietergroenendijk.notifications.strategy.NotificationSendStrategyFactory;
-import com.github.pietergroenendijk.notifications.strategy.WarningNotificationSendStrategy;
+import com.github.pietergroenendijk.storage.notifications.NotificationTaskRepository;
+import com.github.pietergroenendijk.services.notifications.generators.ReturnDateNotificationTaskGenerator;
+import com.github.pietergroenendijk.services.notifications.send_strategies.AlertNotificationSendStrategy;
+import com.github.pietergroenendijk.services.notifications.send_strategies.NotificationSendStrategyFactory;
+import com.github.pietergroenendijk.services.notifications.send_strategies.WarningNotificationSendStrategy;
 
 public class NotificationTaskFactory {
     private final ReturnDateNotificationTaskGenerator RETURN_DATE_TASK_GENERATOR;
