+++ application/src/main/java/com/github/pietergroenendijk/storage/notifications/INotificationTaskRepository.java
@@ -1,4 +1,6 @@
-package com.github.pietergroenendijk.notifications;
+package com.github.pietergroenendijk.storage.notifications;
+
+import com.github.pietergroenendijk.services.notifications.task.NotificationTask;
 
 import java.time.LocalDateTime;
