+++ application/src/main/java/com/github/pietergroenendijk/storage/notifications/NotificationTaskStoreStrategy.java
@@ -1,0 +1,7 @@
+package com.github.pietergroenendijk.storage.notifications;
+
+import com.github.pietergroenendijk.services.notifications.task.NotificationTask;
+
+public interface NotificationTaskStoreStrategy {
+    public void store(NotificationTask task);
+}
